refactor: extract embedded `mod tests` for 2 more god-files (batch 3) (#874)
Final batch of the cross-subsystem "extract embedded mod tests" sweep (after
dml #871 and batch 2 #873). Each file's `#[cfg(test)] mod tests { ... }` moved
to a sibling `<stem>_tests.rs` via `#[cfg(test)] #[path] mod tests;`. Zero
behavioral change; identical test suites.

- storage/engines/core/formats/proximablocks/block_structures.rs: 6,982 -> 5,711 (tests=1,271)
- network/arrow_ipc/service.rs: 3,877 -> 3,117 (tests=760)

Verified: cargo check --tests -p proximadb clean (exit 0); rustfmt 1.95.0 clean.
