+++ src/storage/engines/core/formats/proximablocks/block_structures.rs
@@ -5707,1276 +5707,5 @@ impl Default for BlockMetadataStats {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_data_block_creation() {
-        let records = vec![
-            {
-                let values = vec![1.0f32, 2.0, 3.0];
-                let dim = values.len() as u32;
-                let mut r = ProximaRecord {
-                    oid: "vec_1".to_string(),
-                    created_at_ns: 1_000 * 1_000_000,
-                    ..Default::default()
-                };
-                r.embeddings.push(EmbeddingCell {
-                    model_id: "default".to_string(),
-                    modality: "vector".to_string(),
-                    values: proximadb_records::EmbeddingValues::Fp32(values),
-                    dim,
-                    ..Default::default()
-                });
-                r
-            },
-            {
-                let values = vec![4.0f32, 5.0, 6.0];
-                let dim = values.len() as u32;
-                let mut r = ProximaRecord {
-                    oid: "vec_2".to_string(),
-                    created_at_ns: 2_000 * 1_000_000,
-                    ..Default::default()
-                };
-                r.embeddings.push(EmbeddingCell {
-                    model_id: "default".to_string(),
-                    modality: "vector".to_string(),
-                    values: proximadb_records::EmbeddingValues::Fp32(values),
-                    dim,
-                    ..Default::default()
-                });
-                r
-            },
-        ];
-
-        let compression_config = BlockCompressionConfig::default();
-
-        let block = ProximaDataBlock::new(records, compression_config);
-
-        assert_eq!(block.metadata.record_count, 2);
-        assert_eq!(block.id_range.0, "vec_1");
-        assert_eq!(block.id_range.1, "vec_2");
-        assert_eq!(block.timestamp_range, (1000, 2000));
-    }
-
-    #[test]
-    fn test_superblock_management() {
-        let mut superblock = SuperBlock::new(1, "/path/to/file".to_string());
-
-        let block = ProximaDataBlock::new(
-            vec![ProximaRecord::default()],
-            BlockCompressionConfig::default(),
-        );
-
-        superblock.add_block(block);
-
-        assert_eq!(superblock.blocks.len(), 1);
-        assert_eq!(superblock.record_count, 1);
-    }
-
-    #[test]
-    fn test_grouped_vector_encoding() {
-        // Test GroupedFieldEncodedAndCompressedVector strategy for high-dimensional vectors
-        let dimension = 256; // Should trigger GroupedFieldEncodedAndCompressedVector with Auto
-        let vector_count = 10;
-
-        // Create test vectors
-        let records: Vec<ProximaRecord> = (0..vector_count)
-            .map(|i| {
-                let values: Vec<f32> = (0..dimension)
-                    .map(|d| ((i as f32 * 0.1) + (d as f32 * 0.01)).sin())
-                    .collect();
-                let dim = values.len() as u32;
-                let mut r = ProximaRecord {
-                    oid: format!("vec_{i}"),
-                    ..Default::default()
-                };
-                r.embeddings.push(EmbeddingCell {
-                    model_id: "default".to_string(),
-                    modality: "vector".to_string(),
-                    values: proximadb_records::EmbeddingValues::Fp32(values),
-                    dim,
-                    ..Default::default()
-                });
-                r
-            })
-            .collect();
-
-        // Test Auto strategy (should pick GroupedFieldEncodedAndCompressedVector for D > 128)
-        let compression_config_auto = BlockCompressionConfig {
-            vector_layout: VectorEncodingLayout::Auto,
-            ..Default::default()
-        };
-        let block_auto = ProximaDataBlock::new(records.clone(), compression_config_auto);
-
-        // Serialize and deserialize
-        let serialized = block_auto
-            .serialize()
-            .expect("Auto strategy block should serialize");
-        let deserialized = ProximaDataBlock::deserialize(&serialized, None)
-            .expect("Auto strategy block should deserialize");
-
-        // Verify records match
-        assert_eq!(deserialized.records.len(), vector_count);
-        for (i, record) in deserialized.records.iter().enumerate() {
-            let vec = record
-                .embeddings
-                .first()
-                .map(|e| e.as_fp32_slice())
-                .unwrap_or(&[]);
-            assert_eq!(vec.len(), dimension);
-            // Check first value to ensure correctness
-            let expected = ((i as f32 * 0.1) + (0 as f32 * 0.01)).sin();
-            let diff = (vec[0] - expected).abs();
-            assert!(diff < 0.0001, "Vector mismatch at index {}", i);
-        }
-
-        // Test explicit GroupedFieldEncodedAndCompressedVector strategy
-        let compression_config_grouped = BlockCompressionConfig {
-            vector_layout: VectorEncodingLayout::GroupedFieldEncodedAndCompressedVector,
-            ..Default::default()
-        };
-        let block_grouped = ProximaDataBlock::new(records, compression_config_grouped);
-
-        let serialized_grouped = block_grouped
-            .serialize()
-            .expect("Grouped block should serialize");
-        let deserialized_grouped = ProximaDataBlock::deserialize(&serialized_grouped, None)
-            .expect("Grouped block should deserialize");
-
-        assert_eq!(deserialized_grouped.records.len(), vector_count);
-        // Verify all dimensions are preserved
-        for record in deserialized_grouped.records.iter() {
-            let vec_len = record
-                .embeddings
-                .first()
-                .map(|e| e.values.len())
-                .unwrap_or(0);
-            assert_eq!(vec_len, dimension);
-        }
-    }
-
-    #[test]
-    fn test_block_id_lookup() {
-        let records = vec![{
-            let values = vec![1.0f32, 2.0];
-            let dim = values.len() as u32;
-            let mut r = ProximaRecord {
-                oid: "test_id".to_string(),
-                ..Default::default()
-            };
-            r.embeddings.push(EmbeddingCell {
-                model_id: "default".to_string(),
-                modality: "vector".to_string(),
-                values: proximadb_records::EmbeddingValues::Fp32(values),
-                dim,
-                ..Default::default()
-            });
-            r
-        }];
-
-        let block = ProximaDataBlock::new(records, BlockCompressionConfig::default());
-
-        assert!(block.find_record_by_id("test_id").is_some());
-        assert!(block.find_record_by_id("non_existent").is_none());
-    }
-
-    #[test]
-    fn test_grouped_field_compression_constant_pattern() {
-        // Test Case 1: Constant pattern data that compresses well
-        let dimension = 128;
-        let count = 100;
-
-        // Create constant vectors (all 42.0)
-        let records: Vec<ProximaRecord> = (0..count)
-            .map(|i| {
-                let values = vec![42.0f32; dimension];
-                let dim = values.len() as u32;
-                let mut r = ProximaRecord {
-                    oid: format!("const_{i}"),
-                    ..Default::default()
-                };
-                r.embeddings.push(EmbeddingCell {
-                    model_id: "default".to_string(),
-                    modality: "vector".to_string(),
-                    values: proximadb_records::EmbeddingValues::Fp32(values),
-                    dim,
-                    ..Default::default()
-                });
-                r
-            })
-            .collect();
-
-        let config = BlockCompressionConfig {
-            vector_layout: VectorEncodingLayout::GroupedFieldEncodedAndCompressedVector,
-            algorithm: CompressionAlgorithm::Lz4,
-            compression_level: 1,
-            enable_vector_compression: true,
-            enable_metadata_compression: false,
-            compression_threshold_bytes: 0,
-            dictionary_compression: false,
-            metadata_algorithm: None,
-        };
-
-        let block = ProximaDataBlock::new(records.clone(), config.clone());
-        let serialized = block
-            .serialize_with_config(&config)
-            .expect("Constant pattern block should serialize with config");
-
-        // Verify compression is effective (should be much smaller than raw)
-        let raw_size = count * dimension * 4; // 4 bytes per f32
-        let compression_ratio = raw_size as f64 / serialized.len() as f64;
-        assert!(
-            compression_ratio > 10.0,
-            "Constant data should compress well: {:.2}x",
-            compression_ratio
-        );
-
-        // Verify round-trip
-        let deserialized = ProximaDataBlock::deserialize(&serialized, None)
-            .expect("Constant pattern block should deserialize");
-        assert_eq!(deserialized.records.len(), count);
-
-        // Verify data integrity
-        for (i, record) in deserialized.records.iter().enumerate() {
-            let vec = record
-                .embeddings
-                .first()
-                .map(|e| e.as_fp32_slice())
-                .unwrap_or(&[]);
-            assert_eq!(vec.len(), dimension);
-            for &val in vec {
-                assert!(
-                    (val - 42.0).abs() < 0.0001,
-                    "Record {} has incorrect value",
-                    i
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn test_grouped_field_compression_random_pattern() {
-        // Test Case 2: Random pattern data that doesn't compress well
-        use rand::prelude::*;
-        let dimension = 128;
-        let count = 100;
-        let mut rng = rand::thread_rng();
-
-        // Create random vectors
-        let records: Vec<ProximaRecord> = (0..count)
-            .map(|i| {
-                let values: Vec<f32> = (0..dimension).map(|_| rng.gen_range(-1.0..1.0)).collect();
-                let dim = values.len() as u32;
-                let mut r = ProximaRecord {
-                    oid: format!("random_{i}"),
-                    ..Default::default()
-                };
-                r.embeddings.push(EmbeddingCell {
-                    model_id: "default".to_string(),
-                    modality: "vector".to_string(),
-                    values: proximadb_records::EmbeddingValues::Fp32(values),
-                    dim,
-                    ..Default::default()
-                });
-                r
-            })
-            .collect();
-
-        let config = BlockCompressionConfig {
-            vector_layout: VectorEncodingLayout::GroupedFieldEncodedAndCompressedVector,
-            algorithm: CompressionAlgorithm::Zstd,
-            compression_level: 3,
-            enable_vector_compression: true,
-            enable_metadata_compression: false,
-            compression_threshold_bytes: 0,
-            dictionary_compression: false,
-            metadata_algorithm: None,
-        };
-
-        let block = ProximaDataBlock::new(records.clone(), config.clone());
-        let serialized = block
-            .serialize_with_config(&config)
-            .expect("Random pattern block should serialize with config");
-
-        // Verify compression is less effective (random data doesn't compress well)
-        let raw_size = count * dimension * 4; // 4 bytes per f32
-        let compression_ratio = raw_size as f64 / serialized.len() as f64;
-
-        // NOTE: The ProximaEncoder scheme selection may misidentify random data patterns
-        // (e.g., Small-sample random data may appear "sequential" and get Simple8b encoding)
-        // This can result in unexpectedly high compression ratios for truly random data
-        // Deferred: Improve pattern detection in analyze_and_choose_scheme_f32() to handle random data better
-        // For now, just verify serialization succeeds
-        assert!(
-            compression_ratio > 0.1,
-            "Should produce some output: {:.2}x",
-            compression_ratio
-        );
-
-        // Verify round-trip - handle potential error gracefully
-        match ProximaDataBlock::deserialize(&serialized, None) {
-            Ok(deserialized) => {
-                assert_eq!(
-                    deserialized.records.len(),
-                    count,
-                    "Expected {} records, got {}",
-                    count,
-                    deserialized.records.len()
-                );
-
-                // Verify data integrity
-                for (i, (original, deserialized)) in
-                    records.iter().zip(deserialized.records.iter()).enumerate()
-                {
-                    let orig_vec = original
-                        .embeddings
-                        .first()
-                        .map(|e| e.as_fp32_slice())
-                        .unwrap_or(&[]);
-                    let deser_vec = deserialized
-                        .embeddings
-                        .first()
-                        .map(|e| e.as_fp32_slice())
-                        .unwrap_or(&[]);
-                    assert_eq!(
-                        orig_vec.len(),
-                        deser_vec.len(),
-                        "Record {} dimension mismatch",
-                        i
-                    );
-                    for (j, (&orig, &deser)) in orig_vec.iter().zip(deser_vec.iter()).enumerate() {
-                        assert!(
-                            (orig - deser).abs() < 0.0001,
-                            "Record {} dim {} mismatch: {} vs {}",
-                            i,
-                            j,
-                            orig,
-                            deser
-                        );
-                    }
-                }
-            }
-            Err(e) => {
-                // For random data with aggressive compression, sometimes the compressed data
-                // might not decompress correctly. This is acceptable.
-                println!(
-                    "Random pattern compression test: Deserialization failed (expected for highly random data): {}",
-                    e
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn test_grouped_field_compression_mixed_pattern() {
-        // Test Case 3: Mixed pattern - some groups compress well, others don't
-        let dimension = 128;
-        let count = 100;
-
-        // Create mixed pattern vectors
-        let records: Vec<ProximaRecord> = (0..count)
-            .map(|i| {
-                let vector = if i < count / 2 {
-                    // First half: constant values (compress well)
-                    vec![i as f32; dimension]
-                } else {
-                    // Second half: sequential values (moderate compression)
-                    (0..dimension).map(|d| (i + d) as f32).collect()
-                };
-
-                {
-                    let dim = vector.len() as u32;
-                    let mut r = ProximaRecord {
-                        oid: format!("mixed_{i}"),
-                        ..Default::default()
-                    };
-                    r.embeddings.push(EmbeddingCell {
-                        model_id: "default".to_string(),
-                        modality: "vector".to_string(),
-                        values: proximadb_records::EmbeddingValues::Fp32(vector),
-                        dim,
-                        ..Default::default()
-                    });
-                    r
-                }
-            })
-            .collect();
-
-        let config = BlockCompressionConfig {
-            vector_layout: VectorEncodingLayout::GroupedFieldEncodedAndCompressedVector,
-            algorithm: CompressionAlgorithm::Snappy,
-            compression_level: 1,
-            enable_vector_compression: true,
-            enable_metadata_compression: false,
-            compression_threshold_bytes: 0,
-            dictionary_compression: false,
-            metadata_algorithm: None,
-        };
-
-        let block = ProximaDataBlock::new(records.clone(), config.clone());
-        let serialized = block
-            .serialize_with_config(&config)
-            .expect("Mixed pattern block should serialize with config");
-
-        // Verify moderate compression (between constant and random)
-        let raw_size = count * dimension * 4; // 4 bytes per f32
-        let compression_ratio = raw_size as f64 / serialized.len() as f64;
-        // Mixed pattern should compress moderately (between 1.0x and 20.0x)
-        // ProximaCodec can achieve better compression than old encoder
-        assert!(
-            compression_ratio > 0.5 && compression_ratio < 20.0,
-            "Mixed data should have moderate compression: {:.2}x",
-            compression_ratio
-        );
-
-        // Verify round-trip - handle potential error gracefully
-        match ProximaDataBlock::deserialize(&serialized, None) {
-            Ok(deserialized) => {
-                assert_eq!(deserialized.records.len(), count);
-
-                // Verify data integrity
-                for (i, (original, deserialized)) in
-                    records.iter().zip(deserialized.records.iter()).enumerate()
-                {
-                    let orig_vec = original
-                        .embeddings
-                        .first()
-                        .map(|e| e.as_fp32_slice())
-                        .unwrap_or(&[]);
-                    let deser_vec = deserialized
-                        .embeddings
-                        .first()
-                        .map(|e| e.as_fp32_slice())
-                        .unwrap_or(&[]);
-                    assert_eq!(
-                        orig_vec.len(),
-                        deser_vec.len(),
-                        "Record {} dimension mismatch",
-                        i
-                    );
-                    for (j, (&orig, &deser)) in orig_vec.iter().zip(deser_vec.iter()).enumerate() {
-                        assert!(
-                            (orig - deser).abs() < 0.0001,
-                            "Record {} dim {} mismatch: {} vs {}",
-                            i,
-                            j,
-                            orig,
-                            deser
-                        );
-                    }
-                }
-            }
-            Err(e) => {
-                // For mixed data with varying compression patterns, sometimes issues can occur
-                println!(
-                    "Mixed pattern compression test: Deserialization failed (can happen with mixed patterns): {}",
-                    e
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn test_generate_bloom() {
-        use proximadb_records::{EmbeddingCell, ProximaRecord};
-
-        // Create test records
-        let records = vec![
-            {
-                let values = vec![1.0f32, 2.0, 3.0];
-                let dim = values.len() as u32;
-                let mut r = ProximaRecord {
-                    oid: "vec1".to_string(),
-                    ..Default::default()
-                };
-                r.embeddings.push(EmbeddingCell {
-                    model_id: "default".to_string(),
-                    modality: "vector".to_string(),
-                    values: proximadb_records::EmbeddingValues::Fp32(values),
-                    dim,
-                    ..Default::default()
-                });
-                r
-            },
-            {
-                let values = vec![4.0f32, 5.0, 6.0];
-                let dim = values.len() as u32;
-                let mut r = ProximaRecord {
-                    oid: "vec2".to_string(),
-                    ..Default::default()
-                };
-                r.embeddings.push(EmbeddingCell {
-                    model_id: "default".to_string(),
-                    modality: "vector".to_string(),
-                    values: proximadb_records::EmbeddingValues::Fp32(values),
-                    dim,
-                    ..Default::default()
-                });
-                r
-            },
-            {
-                let values = vec![7.0f32, 8.0, 9.0];
-                let dim = values.len() as u32;
-                let mut r = ProximaRecord {
-                    oid: "vec3".to_string(),
-                    ..Default::default()
-                };
-                r.embeddings.push(EmbeddingCell {
-                    model_id: "default".to_string(),
-                    modality: "vector".to_string(),
-                    values: proximadb_records::EmbeddingValues::Fp32(values),
-                    dim,
-                    ..Default::default()
-                });
-                r
-            },
-        ];
-
-        let block = ProximaDataBlock::new(records.clone(), BlockCompressionConfig::default());
-
-        // Test bloom filter generation
-        let bloom_result = block.generate_bloom();
-        assert!(bloom_result.is_ok());
-
-        let bloom_data = bloom_result.expect("Bloom filter generation should succeed");
-        assert!(bloom_data.is_some());
-
-        let bloom_bytes = bloom_data.expect("Bloom filter data should be present");
-        assert!(!bloom_bytes.is_empty());
-
-        // Verify bloom filter can be deserialized
-        use crate::core::bloom::{BloomFilterConfig, factory::BloomFilterFactory};
-        let config = BloomFilterConfig::for_sstable(records.len());
-        let deserialized = BloomFilterFactory::deserialize(&config, &bloom_bytes);
-        assert!(deserialized.is_ok());
-    }
-
-    #[test]
-    #[allow(clippy::panic)] // Test panic for failure assertion
-    fn test_serialize_with_bloom_sync() {
-        use proximadb_records::{EmbeddingCell, ProximaRecord};
-
-        // Create test records
-        let records = vec![
-            {
-                let values = vec![1.0f32, 2.0];
-                let dim = values.len() as u32;
-                let mut r = ProximaRecord {
-                    oid: "test1".to_string(),
-                    ..Default::default()
-                };
-                r.embeddings.push(EmbeddingCell {
-                    model_id: "default".to_string(),
-                    modality: "vector".to_string(),
-                    values: proximadb_records::EmbeddingValues::Fp32(values),
-                    dim,
-                    ..Default::default()
-                });
-                r
-            },
-            {
-                let values = vec![3.0f32, 4.0];
-                let dim = values.len() as u32;
-                let mut r = ProximaRecord {
-                    oid: "test2".to_string(),
-                    ..Default::default()
-                };
-                r.embeddings.push(EmbeddingCell {
-                    model_id: "default".to_string(),
-                    modality: "vector".to_string(),
-                    values: proximadb_records::EmbeddingValues::Fp32(values),
-                    dim,
-                    ..Default::default()
-                });
-                r
-            },
-        ];
-
-        let block = ProximaDataBlock::new(records, BlockCompressionConfig::default());
-
-        // Test parallel serialization with bloom
-        let result = block.serialize_with_bloom_sync();
-        assert!(result.is_ok());
-
-        let (serialized_block, bloom_data) =
-            result.expect("Serialize with bloom sync should succeed");
-
-        // Verify block was serialized
-        assert!(!serialized_block.is_empty());
-
-        // Verify bloom filter was generated
-        assert!(bloom_data.is_some());
-        assert!(
-            !bloom_data
-                .expect("Bloom filter data should be present")
-                .is_empty()
-        );
-
-        // Verify block can be deserialized
-        let deserialized_block = ProximaDataBlock::deserialize(&serialized_block, None);
-        if let Err(e) = &deserialized_block {
-            panic!("Deserialization failed: {}", e);
-        }
-        assert!(deserialized_block.is_ok());
-        assert_eq!(
-            deserialized_block
-                .expect("Deserialized block should be present")
-                .records
-                .len(),
-            2
-        );
-    }
-
-    #[tokio::test]
-    async fn test_serialize_with_bloom_async() {
-        use proximadb_records::{EmbeddingCell, ProximaRecord};
-
-        // Create test records
-        let values = vec![10.0f32, 20.0, 30.0];
-        let dim = values.len() as u32;
-        let mut async_rec = ProximaRecord {
-            oid: "async1".to_string(),
-            ..Default::default()
-        };
-        async_rec.embeddings.push(EmbeddingCell {
-            model_id: "default".to_string(),
-            modality: "vector".to_string(),
-            values: proximadb_records::EmbeddingValues::Fp32(values),
-            dim,
-            ..Default::default()
-        });
-        let records = vec![async_rec];
-
-        let block = ProximaDataBlock::new(records, BlockCompressionConfig::default());
-
-        // Test async parallel serialization
-        let result = block.serialize_with_bloom().await;
-        assert!(result.is_ok());
-
-        let (serialized_block, bloom_data) =
-            result.expect("Serialize with bloom async should succeed");
-
-        // Verify both were generated
-        assert!(!serialized_block.is_empty());
-        assert!(bloom_data.is_some());
-    }
-
-    #[test]
-    fn test_empty_block_bloom() {
-        // Test with empty records
-        let block = ProximaDataBlock::new(vec![], BlockCompressionConfig::default());
-
-        // Empty block should return None for bloom
-        let bloom_result = block.generate_bloom();
-        assert!(bloom_result.is_ok());
-        assert!(
-            bloom_result
-                .expect("Bloom generation for empty block should succeed")
-                .is_none()
-        );
-
-        // Sync serialization with empty block
-        let result = block.serialize_with_bloom_sync();
-        assert!(result.is_ok());
-        let (_, bloom_data) =
-            result.expect("Serialize with bloom sync for empty block should succeed");
-        assert!(bloom_data.is_none());
-    }
-
-    // ============================================================================
-    // COMPREHENSIVE ENCODING STRATEGY TESTS
-    // ============================================================================
-
-    mod encoding_strategy_tests {
-        use super::*;
-
-        /// Helper to create test vectors with specific patterns
-        fn create_test_vectors(count: usize, dims: usize, pattern: &str) -> Vec<ProximaRecord> {
-            (0..count)
-                .map(|i| {
-                    let vector = match pattern {
-                        "sequential" => (0..dims).map(|d| (i * dims + d) as f32).collect(),
-                        "normalized" => {
-                            let v: Vec<f32> = (0..dims)
-                                .map(|d| ((i as f32 * 0.1) + (d as f32 * 0.01)).sin())
-                                .collect();
-                            let norm = v.iter().map(|x| x * x).sum::<f32>().sqrt();
-                            v.iter().map(|x| x / norm).collect()
-                        }
-                        "constant" => vec![42.0; dims],
-                        "sparse" => {
-                            let mut v = vec![0.0; dims];
-                            v[i % dims] = 1.0;
-                            v
-                        }
-                        "random" => (0..dims)
-                            .map(|d| ((i * 7 + d * 13) % 100) as f32 / 100.0)
-                            .collect(),
-                        _ => vec![0.0; dims],
-                    };
-                    {
-                        let dim = vector.len() as u32;
-                        let mut r = ProximaRecord {
-                            oid: format!("vec_{i}"),
-                            created_at_ns: i as i64 * 1_000_000,
-                            ..Default::default()
-                        };
-                        r.embeddings.push(EmbeddingCell {
-                            model_id: "default".to_string(),
-                            modality: "vector".to_string(),
-                            values: proximadb_records::EmbeddingValues::Fp32(vector),
-                            dim,
-                            ..Default::default()
-                        });
-                        r
-                    }
-                })
-                .collect()
-        }
-
-        /// Verify roundtrip accuracy for encoding/decoding
-        fn verify_roundtrip(original: &[ProximaRecord], decoded: &[ProximaRecord], tolerance: f32) {
-            assert_eq!(original.len(), decoded.len(), "Record count mismatch");
-
-            for (i, (orig, dec)) in original.iter().zip(decoded.iter()).enumerate() {
-                assert_eq!(orig.oid, dec.oid, "ID mismatch at record {}", i);
-                let orig_vec = orig
-                    .embeddings
-                    .first()
-                    .map(|e| e.as_fp32_slice())
-                    .unwrap_or(&[]);
-                let dec_vec = dec
-                    .embeddings
-                    .first()
-                    .map(|e| e.as_fp32_slice())
-                    .unwrap_or(&[]);
-                assert_eq!(
-                    orig_vec.len(),
-                    dec_vec.len(),
-                    "Dimension mismatch at record {}",
-                    i
-                );
-                assert_eq!(
-                    orig.created_at_ns, dec.created_at_ns,
-                    "Timestamp mismatch at record {}",
-                    i
-                );
-
-                for (d, (&orig_val, &dec_val)) in orig_vec.iter().zip(dec_vec.iter()).enumerate() {
-                    let diff = (orig_val - dec_val).abs();
-                    assert!(
-                        diff <= tolerance,
-                        "Vector mismatch at record {} dim {}: expected {}, got {}, diff {}",
-                        i,
-                        d,
-                        orig_val,
-                        dec_val,
-                        diff
-                    );
-                }
-            }
-        }
-
-        // ========================================================================
-        // TransposeFieldEncoded Tests (TV and TB formats)
-        // ========================================================================
-
-        #[test]
-        fn test_transpose_field_encoded_compressed_basic() {
-            // Test TransposeFieldEncodedAndCompressedVector (TV format)
-            // Uses per-dimension encoding: D0=[R0,R1,...], D1=[R0,R1,...]
-
-            let vectors = create_test_vectors(50, 32, "sequential");
-
-            let config = BlockCompressionConfig {
-                vector_layout: VectorEncodingLayout::TransposeFieldEncodedAndCompressedVector,
-                ..Default::default()
-            };
-
-            let block = ProximaDataBlock::new(vectors.clone(), config);
-            let serialized = block
-                .serialize()
-                .expect("Transpose field encoded block should serialize");
-            let deserialized = ProximaDataBlock::deserialize(&serialized, None)
-                .expect("Transpose field encoded block should deserialize");
-
-            verify_roundtrip(&vectors, &deserialized.records, 0.0001);
-        }
-
-        #[test]
-        fn test_transpose_field_encoded_compressed_normalized() {
-            // Test with normalized embeddings (common ML pattern)
-            let vectors = create_test_vectors(100, 128, "normalized");
-
-            let config = BlockCompressionConfig {
-                vector_layout: VectorEncodingLayout::TransposeFieldEncodedAndCompressedVector,
-                ..Default::default()
-            };
-
-            let block = ProximaDataBlock::new(vectors.clone(), config);
-            let serialized = block
-                .serialize()
-                .expect("Transpose field encoded normalized block should serialize");
-            let deserialized = ProximaDataBlock::deserialize(&serialized, None)
-                .expect("Transpose field encoded normalized block should deserialize");
-
-            verify_roundtrip(&vectors, &deserialized.records, 0.0001);
-        }
-
-        #[test]
-        fn test_transpose_field_encoded_compressed_sparse() {
-            // Test with sparse vectors (mostly zeros)
-            // Currently fails with "Unknown scheme marker: 0x01" during deserialization
-            // This appears to be a format mismatch between encoder and decoder
-            let vectors = create_test_vectors(30, 64, "sparse");
-
-            let config = BlockCompressionConfig {
-                vector_layout: VectorEncodingLayout::TransposeFieldEncodedAndCompressedVector,
-                ..Default::default()
-            };
-
-            let block = ProximaDataBlock::new(vectors.clone(), config);
-            let serialized = block
-                .serialize()
-                .expect("Transpose field encoded sparse block should serialize");
-            let deserialized = ProximaDataBlock::deserialize(&serialized, None)
-                .expect("Transpose field encoded sparse block should deserialize");
-
-            verify_roundtrip(&vectors, &deserialized.records, 0.0);
-        }
-
-        #[test]
-        fn test_transpose_field_block_compressed_basic() {
-            // Test TransposeFieldEncodedBlockCompressedVector (TB format)
-            // Uses block-based compression on top of per-dimension encoding
-
-            let vectors = create_test_vectors(50, 32, "random");
-
-            let config = BlockCompressionConfig {
-                vector_layout: VectorEncodingLayout::TransposeFieldEncodedBlockCompressedVector,
-                ..Default::default()
-            };
-
-            let block = ProximaDataBlock::new(vectors.clone(), config);
-            let serialized = block
-                .serialize()
-                .expect("Transpose field block compressed block should serialize");
-            let deserialized = ProximaDataBlock::deserialize(&serialized, None)
-                .expect("Transpose field block compressed block should deserialize");
-
-            verify_roundtrip(&vectors, &deserialized.records, 0.0001);
-        }
-
-        #[test]
-        fn test_transpose_field_block_compressed_high_dim() {
-            // Test with higher dimensions (384 - common for embeddings)
-            let vectors = create_test_vectors(20, 384, "normalized");
-
-            let config = BlockCompressionConfig {
-                vector_layout: VectorEncodingLayout::TransposeFieldEncodedBlockCompressedVector,
-                ..Default::default()
-            };
-
-            let block = ProximaDataBlock::new(vectors.clone(), config);
-            let serialized = block
-                .serialize()
-                .expect("Transpose field block compressed high dim block should serialize");
-            let deserialized = ProximaDataBlock::deserialize(&serialized, None)
-                .expect("Transpose field block compressed high dim block should deserialize");
-
-            verify_roundtrip(&vectors, &deserialized.records, 0.0001);
-        }
-
-        #[test]
-        fn test_transpose_field_constant_values() {
-            // Test with constant values (edge case for encoding)
-            let vectors = create_test_vectors(25, 64, "constant");
-
-            let config = BlockCompressionConfig {
-                vector_layout: VectorEncodingLayout::TransposeFieldEncodedAndCompressedVector,
-                ..Default::default()
-            };
-
-            let block = ProximaDataBlock::new(vectors.clone(), config);
-            let serialized = block
-                .serialize()
-                .expect("Transpose field constant values block should serialize");
-            let deserialized = ProximaDataBlock::deserialize(&serialized, None)
-                .expect("Transpose field constant values block should deserialize");
-
-            verify_roundtrip(&vectors, &deserialized.records, 0.0);
-        }
-
-        // ========================================================================
-        // GroupedFieldEncoded Tests (GV and GB formats)
-        // ========================================================================
-
-        #[test]
-        fn test_grouped_field_encoded_compressed_basic() {
-            // Test GroupedFieldEncodedAndCompressedVector (GV format)
-            // Uses row-wise encoding with 32-dim groups: FG0=[R0[0-31],R1[0-31],...]
-
-            let vectors = create_test_vectors(50, 128, "sequential");
-
-            let config = BlockCompressionConfig {
-                vector_layout: VectorEncodingLayout::GroupedFieldEncodedAndCompressedVector,
-                ..Default::default()
-            };
-
-            let block = ProximaDataBlock::new(vectors.clone(), config);
-            let serialized = block
-                .serialize()
-                .expect("Grouped field encoded compressed block should serialize");
-            let deserialized = ProximaDataBlock::deserialize(&serialized, None)
-                .expect("Grouped field encoded compressed block should deserialize");
-
-            verify_roundtrip(&vectors, &deserialized.records, 0.0001);
-        }
-
-        #[test]
-        fn test_grouped_field_encoded_compressed_256d() {
-            // Test with 256 dimensions (8 groups of 32)
-            let vectors = create_test_vectors(100, 256, "normalized");
-
-            let config = BlockCompressionConfig {
-                vector_layout: VectorEncodingLayout::GroupedFieldEncodedAndCompressedVector,
-                ..Default::default()
-            };
-
-            let block = ProximaDataBlock::new(vectors.clone(), config);
-            let serialized = block
-                .serialize()
-                .expect("Grouped field encoded 256d block should serialize");
-            let deserialized = ProximaDataBlock::deserialize(&serialized, None)
-                .expect("Grouped field encoded 256d block should deserialize");
-
-            verify_roundtrip(&vectors, &deserialized.records, 0.0001);
-        }
-
-        #[test]
-        fn test_grouped_field_encoded_compressed_non_aligned() {
-            // Test with dimensions not multiple of 32 (e.g., 100 dims)
-            let vectors = create_test_vectors(50, 100, "random");
-
-            let config = BlockCompressionConfig {
-                vector_layout: VectorEncodingLayout::GroupedFieldEncodedAndCompressedVector,
-                ..Default::default()
-            };
-
-            let block = ProximaDataBlock::new(vectors.clone(), config);
-            let serialized = block
-                .serialize()
-                .expect("Grouped field encoded non-aligned block should serialize");
-            let deserialized = ProximaDataBlock::deserialize(&serialized, None)
-                .expect("Grouped field encoded non-aligned block should deserialize");
-
-            verify_roundtrip(&vectors, &deserialized.records, 0.0001);
-        }
-
-        #[test]
-        fn test_grouped_field_block_compressed_basic() {
-            // Test GroupedFieldEncodedBlockCompressedVector (GB format)
-            // Uses block compression on top of grouped encoding
-
-            let vectors = create_test_vectors(50, 128, "normalized");
-
-            let config = BlockCompressionConfig {
-                vector_layout: VectorEncodingLayout::GroupedFieldEncodedBlockCompressedVector,
-                ..Default::default()
-            };
-
-            let block = ProximaDataBlock::new(vectors.clone(), config);
-            let serialized = block
-                .serialize()
-                .expect("Grouped field block compressed block should serialize");
-            let deserialized = ProximaDataBlock::deserialize(&serialized, None)
-                .expect("Grouped field block compressed block should deserialize");
-
-            verify_roundtrip(&vectors, &deserialized.records, 0.0001);
-        }
-
-        #[test]
-        fn test_grouped_field_block_compressed_1536d() {
-            // Test with 1536 dimensions (common for OpenAI embeddings)
-            let vectors = create_test_vectors(20, 1536, "normalized");
-
-            let config = BlockCompressionConfig {
-                vector_layout: VectorEncodingLayout::GroupedFieldEncodedBlockCompressedVector,
-                ..Default::default()
-            };
-
-            let block = ProximaDataBlock::new(vectors.clone(), config);
-            let serialized = block
-                .serialize()
-                .expect("Grouped field block compressed 1536d block should serialize");
-            let deserialized = ProximaDataBlock::deserialize(&serialized, None)
-                .expect("Grouped field block compressed 1536d block should deserialize");
-
-            verify_roundtrip(&vectors, &deserialized.records, 0.0001);
-        }
-
-        #[test]
-        fn test_grouped_field_single_group() {
-            // Test with exactly 32 dimensions (single group)
-            let vectors = create_test_vectors(40, 32, "sequential");
-
-            let config = BlockCompressionConfig {
-                vector_layout: VectorEncodingLayout::GroupedFieldEncodedAndCompressedVector,
-                ..Default::default()
-            };
-
-            let block = ProximaDataBlock::new(vectors.clone(), config);
-            let serialized = block
-                .serialize()
-                .expect("Grouped field single group block should serialize");
-            let deserialized = ProximaDataBlock::deserialize(&serialized, None)
-                .expect("Grouped field single group block should deserialize");
-
-            verify_roundtrip(&vectors, &deserialized.records, 0.0001);
-        }
-
-        // ========================================================================
-        // FullVector Tests (planned - not yet implemented)
-        // ========================================================================
-
-        #[test]
-        fn test_full_vector_basic() {
-            // Test FullVector encoding (stores complete vectors)
-            // FV = [R0[all_dims], R1[all_dims], ...]
-
-            let vectors = create_test_vectors(50, 128, "normalized");
-
-            let config = BlockCompressionConfig {
-                vector_layout: VectorEncodingLayout::FullVector,
-                ..Default::default()
-            };
-
-            let block = ProximaDataBlock::new(vectors.clone(), config);
-            let serialized = block
-                .serialize()
-                .expect("Full vector block should serialize");
-            let deserialized = ProximaDataBlock::deserialize(&serialized, None)
-                .expect("Full vector block should deserialize");
-
-            verify_roundtrip(&vectors, &deserialized.records, 0.0001);
-        }
-
-        // ========================================================================
-        // Cross-Strategy Comparison Tests
-        // ========================================================================
-
-        #[test]
-        fn test_compare_transpose_vs_grouped() {
-            // Compare TransposeField vs GroupedField on same data
-            let vectors = create_test_vectors(50, 128, "normalized");
-
-            let transpose_config = BlockCompressionConfig {
-                vector_layout: VectorEncodingLayout::TransposeFieldEncodedAndCompressedVector,
-                ..Default::default()
-            };
-            let transpose_block = ProximaDataBlock::new(vectors.clone(), transpose_config);
-            let transpose_serialized = transpose_block
-                .serialize()
-                .expect("Transpose block should serialize for comparison");
-            let transpose_deserialized = ProximaDataBlock::deserialize(&transpose_serialized, None)
-                .expect("Transpose block should deserialize for comparison");
-
-            let grouped_config = BlockCompressionConfig {
-                vector_layout: VectorEncodingLayout::GroupedFieldEncodedAndCompressedVector,
-                ..Default::default()
-            };
-            let grouped_block = ProximaDataBlock::new(vectors.clone(), grouped_config);
-            let grouped_serialized = grouped_block
-                .serialize()
-                .expect("Grouped block should serialize for comparison");
-            let grouped_deserialized = ProximaDataBlock::deserialize(&grouped_serialized, None)
-                .expect("Grouped block should deserialize for comparison");
-
-            // Both should decode to identical results
-            verify_roundtrip(&vectors, &transpose_deserialized.records, 0.0001);
-            verify_roundtrip(&vectors, &grouped_deserialized.records, 0.0001);
-
-            // Verify both produce same output
-            for (t, g) in transpose_deserialized
-                .records
-                .iter()
-                .zip(grouped_deserialized.records.iter())
-            {
-                let t_vector = t
-                    .embeddings
-                    .first()
-                    .map(|embedding| embedding.as_fp32_slice())
-                    .unwrap_or(&[]);
-                let g_vector = g
-                    .embeddings
-                    .first()
-                    .map(|embedding| embedding.as_fp32_slice())
-                    .unwrap_or(&[]);
-                for (tv, gv) in t_vector.iter().zip(g_vector.iter()) {
-                    assert!(
-                        (tv - gv).abs() < 0.0001,
-                        "Transpose and Grouped produce different results"
-                    );
-                }
-            }
-        }
-
-        #[test]
-        fn test_compression_efficiency() {
-            // Test that encoding provides compression
-            let vectors = create_test_vectors(100, 256, "normalized");
-
-            // Calculate raw size (100 vectors × 256 dims × 4 bytes)
-            let raw_size = vectors.len()
-                * vectors[0]
-                    .embeddings
-                    .first()
-                    .map(|embedding| embedding.values.len())
-                    .unwrap_or(0)
-                * 4;
-
-            // Test GroupedField compression
-            let grouped_config = BlockCompressionConfig {
-                vector_layout: VectorEncodingLayout::GroupedFieldEncodedAndCompressedVector,
-                ..Default::default()
-            };
-            let grouped_block = ProximaDataBlock::new(vectors.clone(), grouped_config);
-            let grouped_serialized = grouped_block
-                .serialize()
-                .expect("Grouped block should serialize for compression efficiency test");
-
-            // Test TransposeField compression
-            let transpose_config = BlockCompressionConfig {
-                vector_layout: VectorEncodingLayout::TransposeFieldEncodedAndCompressedVector,
-                ..Default::default()
-            };
-            let transpose_block = ProximaDataBlock::new(vectors.clone(), transpose_config);
-            let transpose_serialized = transpose_block
-                .serialize()
-                .expect("Transpose block should serialize for compression efficiency test");
-
-            println!("Raw size: {} bytes", raw_size);
-            println!(
-                "Grouped compressed: {} bytes ({:.1}% of raw)",
-                grouped_serialized.len(),
-                (grouped_serialized.len() as f32 / raw_size as f32) * 100.0
-            );
-            println!(
-                "Transpose compressed: {} bytes ({:.1}% of raw)",
-                transpose_serialized.len(),
-                (transpose_serialized.len() as f32 / raw_size as f32) * 100.0
-            );
-
-            // Both should provide some compression (encoded size < raw size)
-            assert!(
-                grouped_serialized.len() < raw_size,
-                "GroupedField should compress data"
-            );
-            assert!(
-                transpose_serialized.len() < raw_size,
-                "TransposeField should compress data"
-            );
-        }
-
-        #[test]
-        fn test_edge_case_single_vector() {
-            // Test with single vector
-            let vectors = create_test_vectors(1, 128, "normalized");
-
-            for layout in [
-                VectorEncodingLayout::TransposeFieldEncodedAndCompressedVector,
-                VectorEncodingLayout::TransposeFieldEncodedBlockCompressedVector,
-                VectorEncodingLayout::GroupedFieldEncodedAndCompressedVector,
-                VectorEncodingLayout::GroupedFieldEncodedBlockCompressedVector,
-            ] {
-                let config = BlockCompressionConfig {
-                    vector_layout: layout,
-                    ..Default::default()
-                };
-
-                let block = ProximaDataBlock::new(vectors.clone(), config);
-                let serialized = block
-                    .serialize()
-                    .expect("Single vector block should serialize");
-                let deserialized = ProximaDataBlock::deserialize(&serialized, None)
-                    .expect("Single vector block should deserialize");
-
-                verify_roundtrip(&vectors, &deserialized.records, 0.0001);
-            }
-        }
-
-        #[test]
-        fn test_edge_case_small_dimension() {
-            // Test with very small dimensions (< 32)
-            let vectors = create_test_vectors(50, 8, "sequential");
-
-            for layout in [
-                VectorEncodingLayout::TransposeFieldEncodedAndCompressedVector,
-                VectorEncodingLayout::GroupedFieldEncodedAndCompressedVector,
-            ] {
-                let config = BlockCompressionConfig {
-                    vector_layout: layout,
-                    ..Default::default()
-                };
-
-                let block = ProximaDataBlock::new(vectors.clone(), config);
-                let serialized = block
-                    .serialize()
-                    .expect("Small dimension block should serialize");
-                let deserialized = ProximaDataBlock::deserialize(&serialized, None)
-                    .expect("Small dimension block should deserialize");
-
-                verify_roundtrip(&vectors, &deserialized.records, 0.0001);
-            }
-        }
-
-        #[test]
-        fn test_large_batch() {
-            // Test with large batch (1000 vectors)
-            let vectors = create_test_vectors(1000, 128, "random");
-
-            let config = BlockCompressionConfig {
-                vector_layout: VectorEncodingLayout::GroupedFieldEncodedAndCompressedVector,
-                ..Default::default()
-            };
-
-            let block = ProximaDataBlock::new(vectors.clone(), config);
-            let serialized = block
-                .serialize()
-                .expect("Large batch block should serialize");
-            let deserialized = ProximaDataBlock::deserialize(&serialized, None)
-                .expect("Large batch block should deserialize");
-
-            verify_roundtrip(&vectors, &deserialized.records, 0.0001);
-        }
-
-        #[test]
-        fn test_lossless_encoding() {
-            // Verify encoding is truly lossless (no quantization)
-            let vectors = create_test_vectors(50, 128, "random");
-
-            for layout in [
-                VectorEncodingLayout::TransposeFieldEncodedAndCompressedVector,
-                VectorEncodingLayout::TransposeFieldEncodedBlockCompressedVector,
-                VectorEncodingLayout::GroupedFieldEncodedAndCompressedVector,
-                VectorEncodingLayout::GroupedFieldEncodedBlockCompressedVector,
-            ] {
-                let config = BlockCompressionConfig {
-                    vector_layout: layout,
-                    ..Default::default()
-                };
-
-                let block = ProximaDataBlock::new(vectors.clone(), config);
-                let serialized = block
-                    .serialize()
-                    .expect("Lossless encoding block should serialize");
-                let deserialized = ProximaDataBlock::deserialize(&serialized, None)
-                    .expect("Lossless encoding block should deserialize");
-
-                // Use very tight tolerance to verify lossless encoding
-                verify_roundtrip(&vectors, &deserialized.records, 1e-6);
-            }
-        }
-    }
-}
+#[path = "block_structures_tests.rs"]
+mod tests;
